+++ src/commands.ts
@@ -89,8 +89,8 @@ import {
   findBackmatter,
   findBibliographySection,
   findBody,
-  insertAttachmentsNode,
   findFootnotesSection,
+  insertAttachmentsNode,
   insertAwardsNode,
   insertFootnotesSection,
   insertSupplementsNode,
