+++ src/components/references/ReferenceForm/config.ts
@@ -1,0 +1,323 @@
+import { BibliographyItemType } from "@manuscripts/transform";
+
+export const bibliographyItemTypes: [BibliographyItemType, string][] = [
+  ['article-journal', 'Journal Article'],
+  ['book', 'Book'],
+  ['chapter', 'Chapter'],
+  ['paper-conference', 'Conference Paper'],
+  ['thesis', 'Thesis'],
+  ['webpage', 'Web Page'],
+  ['other', 'Other'],
+  ['standard', 'Standard'],
+  ['dataset', 'Dataset'],
+  ['preprint', 'Preprint'],
+]
+
+interface FieldConfigObject {
+  [fieldName: string]: {
+    [key in BibliographyItemType]: boolean;
+  };
+}
+
+export const fieldConfigObject: FieldConfigObject = {
+  author: {
+    'article-journal': true,
+    book: true,
+    chapter: true,
+    'paper-conference': true,
+    dataset: true,
+    preprint: true,
+    standard: false,
+    thesis: true,
+    webpage: true,
+    other: true,
+  },
+  editor: {
+    'article-journal': false,
+    book: true,
+    chapter: true,
+    'paper-conference': false,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  title: {
+    'article-journal': true,
+    book: true,
+    chapter: true,
+    'paper-conference': true,
+    dataset: false,
+    preprint: true,
+    standard: false,
+    thesis: true,
+    webpage: true,
+    other: true,
+  },
+  'data-title': {
+    'article-journal': false,
+    book: false,
+    chapter: false,
+    'paper-conference': false,
+    dataset: true,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  std: {
+    'article-journal': false,
+    book: false,
+    chapter: false,
+    'paper-conference': false,
+    dataset: false,
+    preprint: false,
+    standard: true,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  issued: {
+    'article-journal': true,
+    book: true,
+    chapter: true,
+    'paper-conference': true,
+    dataset: true,
+    preprint: true,
+    standard: false,
+    thesis: true,
+    webpage: true,
+    other: true,
+  },
+  source: {
+    'article-journal': true,
+    book: true,
+    chapter: true,
+    'paper-conference': false,
+    dataset: true,
+    preprint: true,
+    standard: true,
+    thesis: false,
+    webpage: false,
+    other: true,
+  },
+  series: {
+    'article-journal': false,
+    book: true,
+    chapter: true,
+    'paper-conference': false,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  volume: {
+    'article-journal': true,
+    book: true,
+    chapter: true,
+    'paper-conference': false,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  issue: {
+    'article-journal': true,
+    book: false,
+    chapter: false,
+    'paper-conference': false,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  supplement: {
+    'article-journal': true,
+    book: false,
+    chapter: false,
+    'paper-conference': false,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  edition: {
+    'article-journal': false,
+    book: true,
+    chapter: true,
+    'paper-conference': false,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  'publisher-place': {
+    'article-journal': false,
+    book: true,
+    chapter: true,
+    'paper-conference': true,
+    dataset: true,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: true,
+    other: true,
+  },
+  'publisher': {
+    'article-journal': false,
+    book: true,
+    chapter: true,
+    'paper-conference': true,
+    dataset: true,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: true,
+    other: true,
+  },
+  event: {
+    'article-journal': false,
+    book: false,
+    chapter: false,
+    'paper-conference': true,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  'event-date': {
+    'article-journal': false,
+    book: false,
+    chapter: false,
+    'paper-conference': true,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  'event-place': {
+    'article-journal': false,
+    book: false,
+    chapter: false,
+    'paper-conference': true,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  institution: {
+    'article-journal': false,
+    book: false,
+    chapter: false,
+    'paper-conference': false,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: true,
+    webpage: false,
+    other: false,
+  },
+  page: {
+    'article-journal': true,
+    book: true,
+    chapter: true,
+    'paper-conference': true,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  'elocationID': {
+    'article-journal': true,
+    book: false,
+    chapter: false,
+    'paper-conference': false,
+    dataset: false,
+    preprint: true,
+    standard: false,
+    thesis: false,
+    webpage: false,
+    other: false,
+  },
+  'links': { // type: ext-link
+    'article-journal': true,
+    book: true,
+    chapter: true,
+    'paper-conference': true,
+    dataset: true,
+    preprint: true,
+    standard: true,
+    thesis: true,
+    webpage: true,
+    other: true,
+  },
+  'number-of-pages': {
+    'article-journal': false,
+    book: true,
+    chapter: true,
+    'paper-conference': false,
+    dataset: false,
+    preprint: false,
+    standard: false,
+    thesis: true,
+    webpage: false,
+    other: false,
+  },
+  'date-in-citation': {
+    'article-journal': false,
+    book: false,
+    chapter: false,
+    'paper-conference': false,
+    dataset: true,
+    preprint: false,
+    standard: false,
+    thesis: false,
+    webpage: true,
+    other: false,
+  },
+  comment: {
+    'article-journal': true,
+    book: true,
+    chapter: true,
+    'paper-conference': false,
+    dataset: true,
+    preprint: false,
+    standard: false,
+    thesis: true,
+    webpage: false,
+    other: true,
+  },
+  'pubIds': { //type: doi, etc...
+    'article-journal': true,
+    book: true,
+    chapter: true,
+    'paper-conference': true,
+    dataset: true,
+    preprint: true,
+    standard: false,
+    thesis: true,
+    webpage: false,
+    other: false,
+  },
+}
